requirements.txt
keystoneauth1==2.1.0
openstacksdk>=0.7.1
python-keystoneclient>=2.0.0
